requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
Levenshtein==0.26.1
openai==1.52.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-dotenv==1.0.1
RapidFuzz==3.10.1
requests==2.32.3
sniffio==1.3.1
SpeechRecognition==3.11.0
srt==3.5.3
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
vosk==0.3.44
websockets==13.1
